Initiating JavaScript at Medac.
- holamundo.html
This file is to show and test the code generated in the js folder.

- js (folder)

- holamundo.js
In this file we are trying which are the variables, how to manipulate then and work with them.

- practicas_iniciales.js
The objective of this file is to do the classes exercises that are to try the conditionals, loops...